+++ src/Temporalio/Client/Interceptors/RestoreOriginalActivityOptionsInput.cs
@@ -1,0 +1,18 @@
+namespace Temporalio.Client.Interceptors
+{
+    /// <summary>
+    /// Input for <see cref="ClientOutboundInterceptor.RestoreOriginalActivityOptionsAsync" />.
+    /// </summary>
+    /// <param name="Id">Activity ID.</param>
+    /// <param name="RunId">Activity run ID if any.</param>
+    /// <param name="RpcOptions">RPC options for the call.</param>
+    /// <remarks>WARNING: Standalone activities are experimental.</remarks>
+    /// <remarks>
+    /// WARNING: This constructor may have required properties added. Do not rely on the exact
+    /// constructor, only use "with" clauses.
+    /// </remarks>
+    public record RestoreOriginalActivityOptionsInput(
+        string Id,
+        string? RunId,
+        RpcOptions? RpcOptions);
+}